Consolidate development samples

+++ samples/nomad/development/initialize.sh
@@ -3,11 +3,26 @@
 set -euo pipefail
 
 curl -Ls https://gist.github.com/gusztavvargadr/1f0d7dddc7f48549368eaaedf19bfe55/raw/provision.sh | sudo CHEF_POLICY="gusztavvargadr_development" bash -s
+sudo apt install -y git jq net-tools
 
-sudo apt remove -y nomad
+CONSUL_VERSION=1.19.2
+if [[ "$(consul --version)" != *"$CONSUL_VERSION"* ]]; then
+  sudo apt-mark unhold consul
+  sudo apt remove -y consul
+  sudo apt install -y consul="$CONSUL_VERSION*"
+  sudo apt-mark hold consul
 
-sudo apt install -y git jq net-tools nomad=1.8.4-1
-sudo apt autoremove -y
-sudo apt clean -y
+  sudo sed -i "s/Type=notify/Type=simple/g" /lib/systemd/system/consul.service
+  sudo systemctl daemon-reload
+fi
 
-sudo apt-mark hold nomad
+NOMAD_VERSION=1.8.4
+if [[ "$(nomad --version)" != *"$NOMAD_VERSION"* ]]; then
+  sudo apt-mark unhold nomad
+  sudo apt remove -y nomad
+  sudo apt install -y nomad="$NOMAD_VERSION*"
+  sudo apt-mark hold nomad
+
+  sudo sed -i "s/Type=notify/Type=simple/g" /lib/systemd/system/nomad.service
+  sudo systemctl daemon-reload
+fi
